Move tests from extenstion to local package
The tests we move here are related strictly to the `local` package and
it seems a better place to have them here.
